Lab 2 part 2
Provide the Javascript ES6 to implement a 1D, 2-colour cellular automaton based on the following description and steps:


Generate a row having 101 cells across made from <div> tags of size 8px by 8px. Each cell will have a randomly initialised state of ‘active’ or ‘inactive’. If active, give it one fill color of your choice. If inactive give it another fill color of your choice
Generate the next row of cells based on the state of the ancestors in the previous row according to the following rule

For example, if the direct ancestor is active and the direct ancestor’s left sibling is active and the direct ancestor’s right sibling is active, then make the corresponding cell in the new row active.


The first cell in a row doesn’t have a left sibling to use the row’s last cell’s state in this case


The last cell in a row doesn’t have a right sibling to use the row’s first cell’s state in this case


Repeat the process until you have created 50 generations